Fix: Correct critical import paths and refactor text rendering utilities
This commit addresses numerous "Failed to resolve import" errors across various components and initiates a refactoring of text rendering helper functions to improve maintainability and consistency.

**Key Changes and Fixes:**

1.  **Critical Import Path Corrections:**
    *   Addressed specific import path errors from user-provided logs for `glossaryData.js`, `textProcessing.js`, and `MermaidDiagram.jsx` in the following files:
        *   `src/pages/GlossaryPage.jsx`
        *   `src/components/caches/FundamentalsView.jsx`
        *   `src/components/caches/CachepediaView.jsx`
        *   `src/components/databases/SectionSqlDB.jsx`
        *   `src/components/databases/SectionReplicationDB.jsx`
        *   `src/components/messaging_queues/DeepDiveModuleMQ.jsx`
        *   `src/components/loadbalancing/AlgorithmsView.jsx`
        *   `src/components/loadbalancing/ScenariosView.jsx`
        *   `src/components/apidesign/ProtocolsView.jsx`
        *   `src/components/apidesign/PatternsView.jsx`
        *   `src/components/apidesign/SecurityView.jsx`
        *   `src/components/apidesign/ScenariosView.jsx`
    *   Systematically reviewed and corrected paths in other `FundamentalsView` components (`databases`, `scalabilityconcepts`) to use `../../` instead of `../../../` for utils and data, and added `.js`/`.jsx` extensions.

2.  **Text Rendering Refactoring (In Progress):**
    *   Created a new utility file `src/utils/textRenderUtils.jsx`.
    *   This file contains new centralized components: `RenderGlossaryLink`, `RenderMarkdownLink`, and `RenderTextWithLinks` (which handles both glossary and markdown links).
    *   Updated several key components to use `RenderTextWithLinks`, removing their local/duplicated text rendering logic. Components updated include:
        *   `caches/FundamentalsView.jsx`
        *   `databases/FundamentalsView.jsx`
        *   `loadbalancing/FundamentalsView.jsx`
        *   `apidesign/FundamentalsView.jsx`
        *   `scalabilityconcepts/FundamentalsView.jsx`
        *   `interviewapproach/InterviewApproachPage.jsx`
        *   `pages/CaseStudyDetailPage.jsx`
        *   `databases/SectionSqlDB.jsx` (partially, due to `dangerouslySetInnerHTML`)
        *   `loadbalancing/AlgorithmsView.jsx`
    *   This refactoring is ongoing; more view components still need to be updated to use these centralized utilities.

3.  **Dependency Management:**
    *   Added `recharts` to `package.json` as it was a missing dependency causing an import error in `TradeoffRadar.jsx`. User has been prompted to run `npm install`.

4.  **Other Previous Fixes (Included in this branch's history):**
    *   Made Cache Eviction Simulator visible.
    *   Ensured Markdown links were clickable on Interview Approach and Case Study Detail pages (now superseded by `RenderTextWithLinks`).
    *   Added comprehensive meta tags to numerous pages.

**Current Status & Next Steps (Interrupted):**
I was in the middle of the "Systematic Import Path Verification (Comprehensive)" and "Refactor Helper Text Processing Functions (Finalize)" steps. The immediate plan was to continue updating the remaining view components (e.g., `CachepediaView.jsx`, other views in `loadbalancing`, `apidesign`, etc.) to use the new centralized `RenderTextWithLinks` utility and ensure all their import paths are correct. After that, the plan was to add basic rendering tests.

This submission includes the progress made so far in stabilizing the application by fixing critical import errors and centralizing text processing logic.
