+++ system-design-study-app/src/components/loadbalancing/ScenariosView.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Typography, Box, Paper, List, ListItem, ListItemText, Divider } from '@mui/material';
 import { Link as RouterLink } from 'react-router-dom';
-import { parseTextForGlossaryLinks, getDefinitionSnippet } from '../../../utils/textProcessing';
-import { glossaryData } from '../../../data/glossaryData';
+import { parseTextForGlossaryLinks, getDefinitionSnippet } from '../../utils/textProcessing.js';
+import { glossaryData } from '../../data/glossaryData.js';
 
 // Helper component to render processed text (strings and links)
 const RenderProcessedText = ({ textParts, component = "span" }) => {
